+++ sessioning/lib/sessioning.middleware.ts
@@ -1,17 +1,27 @@
-import type { BaseRequestContext, Middleware } from "@potami/core";
-import { cookies } from "../deps.ts";
-import type {
-  ISessionStore,
-  SessionFetchOptions,
-} from "./session-store.interface.ts";
-import type {
-  Session,
-  SessionContext,
-  SessionDataSetter,
-} from "./session.model.ts";
+import type { Context, Middleware } from '@potami/core';
+import { cookies } from '../deps.ts';
+import type { ISessionStore, SessionFetchOptions } from './session-store.interface.ts';
+import type { Session, SessionContext, SessionDataSetter } from './session.model.ts';
 
 export interface HandleSessioningOptions<T> {
   store: ISessionStore<T>;
+  /**
+   * The `Context` instance to use for the session context. You'll use this same instance
+   * with `getContext` to read the session data. If you'd like to update the session data,
+   * use the `setSessionData` available on this context instead of calling `setContext`.
+   * 
+   * @example
+   * ```ts
+   * // Do:
+   * const { setSessionData } = getContext({ sessionContext });
+   * 
+   * setSessionData({ myData: 123 });
+   * 
+   * // DON'T:
+   * setContext(sessionContext, { ...getContext(sessionContext), session: { myData: 123 }});
+   * ```
+   */
+  sessionContext: Context<SessionContext<T>>;
   /**
    * The name to use for the session cookie. According
    * to [OWASP recommendations](https://cheatsheetseries.owasp.org/cheatsheets/Session_Management_Cheat_Sheet.html)
@@ -69,7 +79,7 @@ export interface HandleSessioningOptions<T> {
      *
      * Defaults to `'Strict'`.
      */
-    sameSite?: "Strict" | "Lax" | "None";
+    sameSite?: 'Strict' | 'Lax' | 'None';
     /**
      * The domain this cookie is valid for. Avoid setting an overly permissive
      * value. For details, refer to this [OWASP cheatsheet](https://cheatsheetseries.owasp.org/cheatsheets/Session_Management_Cheat_Sheet.html#domain-and-path-attributes).
@@ -114,13 +124,13 @@ export interface HandleSessioningOptions<T> {
  * to handle sessions. Sessions are persisted by leveraging the provided `store`.
  *
  * This middleware handles all aspects of sessioning, including:
- * - Reading the session token off incoming requests
- * - Populating the `ctx.session` based on the session token present or absent on incoming requests
- * - Setting the session token cookie based on the currently valid session
- * - Monitoring for probing by validating incoming session tokens
+ * - Reading the session token off incoming requests.
+ * - Populating the `sessionContext` based on the session token present or absent on incoming requests.
+ * - Setting the session token cookie based on the currently valid session.
+ * - Monitoring for probing by validating incoming session tokens.
  *
- * This middleware guarantees that there is always a `ctx.session` so long as this middleware
- * runs before anything trying to access `ctx.session`. If a request enters the server
+ * This middleware guarantees that there is always a session on the `sessionContext` so long as this middleware
+ * runs before anything that tries to access the `sessionContext`. If a request enters the server
  * without a session token, a new session is created.
  *
  * Sessions allow for app-specific `data` to be added to them. See the documentation for {@link SessionContext.session}
@@ -136,17 +146,15 @@ export interface HandleSessioningOptions<T> {
  * else.
  */
 export const handleSessioning =
-  <
-    SessionDataType,
-    AppContextType extends BaseRequestContext & SessionContext<SessionDataType>,
-  >({
+  <SessionDataType>({
     store,
-    sessionCookieName = "id",
+    sessionContext,
+    sessionCookieName = 'id',
     onReceivedInvalidSessionId,
     sessionFetchOptions,
     cookieAttributes,
-  }: HandleSessioningOptions<SessionDataType>): Middleware<AppContextType> =>
-  async ({ req, resHeaders, ctx }) => {
+  }: HandleSessioningOptions<SessionDataType>): Middleware =>
+  async ({ req, resHeaders, getContext, setContext }) => {
     const sessionId = getSessionIdFromRequest(req, sessionCookieName);
 
     if (sessionId && !store.isSessionIdValid(sessionId)) {
@@ -154,57 +162,48 @@ export const handleSessioning =
     }
 
     const session = sessionId
-      ? (await store.fetchSession(sessionId, sessionFetchOptions)) ??
-        (await store.createSession())
+      ? (await store.fetchSession(sessionId, sessionFetchOptions)) ?? (await store.createSession())
       : await store.createSession();
 
-    setSessionCookieHeader(
-      resHeaders,
-      sessionCookieName,
+    setSessionCookieHeader(resHeaders, sessionCookieName, session, cookieAttributes);
+
+    setContext(sessionContext, {
       session,
-      cookieAttributes,
-    );
+      setSessionData: async (dataOrSetter) => {
+        const updatedSession = await store.setSessionData(
+          session.id,
+          // @ts-ignore - Complaining that it doesn't know which overload to use, which doesn't matter.
+          dataOrSetter
+        );
 
-    ctx.session = session;
-    ctx.setSessionData = async (dataOrSetter) => {
-      const updatedSession = await store.setSessionData(
-        session.id,
-        // @ts-ignore - Complaining that it doesn't know which overload to use, which doesn't matter.
-        dataOrSetter,
-      );
+        let newSession: Session<SessionDataType>;
 
-      let newSession: Session<SessionDataType>;
+        if (updatedSession) {
+          newSession = updatedSession;
+        } else {
+          // Edge case where session was NOT expired when it entered the server,
+          // but is now expired when this function is called. We create a new session
+          // and attach the original session's data to it, effectively extending the
+          // session.
+          const newData: SessionDataType | undefined =
+            typeof dataOrSetter === 'function'
+              ? (dataOrSetter as SessionDataSetter<SessionDataType>)(session.data)
+              : dataOrSetter;
 
-      if (updatedSession) {
-        newSession = updatedSession;
-      } else {
-        // Edge case where session was NOT expired when it entered the server,
-        // but is now expired when this function is called. We create a new session
-        // and attach the original session's data to it, effectively extending the
-        // session.
-        const newData: SessionDataType | undefined =
-          typeof dataOrSetter === "function"
-            ? (dataOrSetter as SessionDataSetter<SessionDataType>)(session.data)
-            : dataOrSetter;
+          newSession = await store.createSession(newData);
+        }
 
-        newSession = await store.createSession(newData);
-      }
+        setSessionCookieHeader(resHeaders, sessionCookieName, newSession, cookieAttributes);
 
-      setSessionCookieHeader(
-        resHeaders,
-        sessionCookieName,
-        newSession,
-        cookieAttributes,
-      );
-
-      ctx.session = newSession;
-    };
+        setContext(sessionContext, {
+          ...getContext(sessionContext),
+          session: newSession,
+        });
+      },
+    });
   };
 
-function getSessionIdFromRequest(
-  req: Request,
-  sessionCookieName: string,
-): string | undefined {
+function getSessionIdFromRequest(req: Request, sessionCookieName: string): string | undefined {
   const readCookies = cookies.getCookies(req.headers);
 
   return readCookies[sessionCookieName];
@@ -217,12 +216,12 @@ function setSessionCookieHeader<T>(
   {
     secure = true,
     httpOnly = true,
-    sameSite = "Strict",
+    sameSite = 'Strict',
     domain,
-    path = "/",
+    path = '/',
     maxAge,
     expires,
-  }: HandleSessioningOptions<T>["cookieAttributes"] = {},
+  }: HandleSessioningOptions<T>['cookieAttributes'] = {}
 ): void {
   const cookie: cookies.Cookie = {
     name: sessionCookieName,
@@ -242,7 +241,7 @@ function setSessionCookieHeader<T>(
     .filter((cookie) => !cookie.includes(`${sessionCookieName}=`))
     .filter(Boolean);
 
-  headers.set("Set-Cookie", cookiesWithoutExistingSessionCookie.join(", "));
+  headers.set('Set-Cookie', cookiesWithoutExistingSessionCookie.join(', '));
 
   cookies.setCookie(headers, cookie);
 }