+++ testing/changelog.md
@@ -8,8 +8,8 @@
   - `makeMiddlewareSubjects`:
     - No longer accepts nor returns a `ctx` object.
     - Now accepts `getContext` and `setContext` functions as properties of its arguments.
-      - If not provided, `getContext` defaults to a function that always gives the provided context's default value.
-      - If not provided, `setContext` defaults to a function that does nothing.
+      - If not provided, `getContext` defaults to a function that will retrieve the value set to the provided context within the scope of this one middleware.
+      - If not provided, `setContext` defaults to a function that will set the value of the context to the provided value within the scope of this one middleware.
 
 # 0.2.0